[fixed] Character deletion issue when name contains an apostrophe [fixed] razor knuckles being named wrong so vibro knucklers should appear on the gungan again (karas) [http://trac2.assembla.com/swgemu/ticket/252 trac ticket 252] [fixed] rebel npc secondaries so they should regen correctly (karas) [fixed] Faction point overflow (mizath) [http://www.swgemu.com/bugs/view.php?id=1295 Mantis Report 1295] made missionmanager error messages more detailed, to track down missionObject null cases
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1479 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/player/faction/FactionPointsMap.h
@@ -65,7 +65,7 @@ public:
 	}
 
 
-	int16 getFactionPoints(String faction) {
+	int32 getFactionPoints(String faction) {
 		if (!containsKey(faction))
 			return 0;
 		else
